@@ -17,7 +17,7 @@
     },
     "require": {
         "php": ">=5.5.9",
-        "symfony/symfony": "3.0.*",
+        "symfony/symfony": "3.2.*",
         "doctrine/orm": "^2.5",
         "doctrine/doctrine-bundle": "^1.6",
         "doctrine/doctrine-cache-bundle": "^1.2",
@@ -28,18 +28,52 @@
         "incenteev/composer-parameter-handler": "^2.0",
         "stof/doctrine-extensions-bundle" : "dev-master",
         "symfony/translation": "^3.0",
-        "friendsofsymfony/user-bundle": "~2.0@dev",
-        "misd/guzzle-bundle": "dev-master"
+        "misd/guzzle-bundle": "dev-master",
+        "predis/predis": "1.0.*",
+        "snc/redis-bundle": "2.x-dev",
+        "jms-serializer/serializer-bundle": "^1.1",
+        "curl/curl": "^1.3",
+        "friendsofsymfony/rest-bundle": "^1.7",
+        "jms/serializer-bundle": "^1.1",
+        "ramsey/uuid": "^3.4",
+        "symfony/security-acl": "^3.0",
+        "friendsofsymfony/elastica-bundle": "^3.2",
+        "friendsofsymfony/jsrouting-bundle": "^1.6",
+        "sg/datatablesbundle": "^0.13.0",
+        "oni-sys/core-bundle": "dev-develop",
+        "oni-sys/user-manager-bundle": "dev-develop",
+        "oni-sys/product-manager-bundle": "dev-develop",
+        "oni-sys/content-manager-bundle": "dev-develop",
+        "oni-sys/dashboard-manager-bundle": "dev-develop"
     },
     "require-dev": {
         "sensio/generator-bundle": "^3.0",
-        "symfony/phpunit-bridge": "^2.7",
         "doctrine/doctrine-fixtures-bundle": "^2.3"
     },
     "repositories": [
         {
             "type": "vcs",
             "url": "https://github.com/shinsantiger/guzzle-bundle"
+        },
+        {
+            "type": "vcs",
+            "url": "https://github.com/peteratkins85/CoreBundle.git"
+        },
+        {
+            "type": "vcs",
+            "url": "https://github.com/peteratkins85/UserManagerBundle.git"
+        },
+        {
+            "type": "vcs",
+            "url": "https://github.com/peteratkins85/ProductManagerBundle.git"
+        },
+        {
+            "type": "vcs",
+            "url": "https://github.com/peteratkins85/ContentManagerBundle.git"
+        },
+        {
+            "type": "vcs",
+            "url": "https://github.com/peteratkins85/DashboardManagerBundle.git"
         }
     ],
     "scripts": {
