@@ -1,32 +1,33 @@
 <?php
-
-use Symfony\Component\HttpFoundation\Request;
+error_reporting(-1);
+ini_set('display_errors', 'on');
+use symfony\Component\Httpfoundation\Request;
 
 /**
- * @var Composer\Autoload\ClassLoader
+ * @var composer\Autoload\Classloader
  */
-$loader = require __DIR__.'/../app/autoload.php';
-include_once __DIR__.'/../var/bootstrap.php.cache';
+$loader = require __dir__.'/../app/autoload.php';
+include_once __dir__.'/../var/bootstrap.php.cache';
 
-// Enable APC for autoloading to improve performance.
-// You should change the ApcClassLoader first argument to a unique prefix
+// enable apc for autoloading to improve performance.
+// you should change the apcclassloader first argument to a unique prefix
 // in order to prevent cache key conflicts with other applications
-// also using APC.
+// also using apc.
 /*
-$apcLoader = new Symfony\Component\ClassLoader\ApcClassLoader(sha1(__FILE__), $loader);
+$apcloader = new symfony\Component\Classloader\Apcclassloader(sha1(__file__), $loader);
 $loader->unregister();
-$apcLoader->register(true);
+$apcloader->register(true);
 */
 
-//require_once __DIR__.'/../app/AppCache.php';
+//require_once __dir__.'/../app/appcache.php';
 
-$kernel = new AppKernel('prod', false);
-$kernel->loadClassCache();
-//$kernel = new AppCache($kernel);
+$kernel = new appkernel('prod', false);
+$kernel->loadclasscache();
+//$kernel = new appcache($kernel);
 
-// When using the HttpCache, you need to call the method in your front controller instead of relying on the configuration parameter
-//Request::enableHttpMethodParameterOverride();
-$request = Request::createFromGlobals();
+// when using the httpcache, you need to call the method in your front controller instead of relying on the configuration parameter
+//request::enablehttpmethodparameteroverride();
+$request = request::createfromglobals();
 $response = $kernel->handle($request);
 $response->send();
 $kernel->terminate($request, $response);
